scripts/rodobens/pdf2md/requirements.txt
# Runtime minimo deterministico
pdfplumber>=0.11,<1
pypdf>=5,<6

# Opcional (instalar apenas se for usar OCR)
Pillow>=10,<12

# Execucao ONNX (opcional, para OCR acelerado)
onnxruntime>=1.18,<2
# Instale este extra apenas quando precisar acelerar em Intel:
# onnxruntime-openvino>=1.18,<2

# Qualidade/infra
PyYAML>=6,<7
rich>=13,<14
